+++ arangod/Cluster/ClusterInfo.cpp
@@ -77,7 +77,7 @@ static inline arangodb::AgencyOperation CreateCollectionOrder(std::string const&
 #endif
   arangodb::AgencyOperation op{"Plan/Collections/" + dbName + "/" + collection,
                                arangodb::AgencyValueOperationType::SET, info};
-  op._ttl = timeout;
+  op._ttl = static_cast<uint64_t>(timeout);
   return op;
 }
